configuring user routes and JWT authentication

+++ Backend/controllerss/goalControllers.ts
@@ -1,16 +1,21 @@
 import { Request, Response } from "express"
 import asyncHandler from "express-async-handler"
 import GoalModel from "../models/goalModel"
+import userModel from "../models/userModel"
 
+
+interface CustomRequest extends Request {
+    user?: any
+}
 /*
     @desc - Get all goals
     @route - GET /api/goals
     @access - Private
 
 */
-export const getGoals = asyncHandler(async (req: Request, res: Response) => {
+export const getGoals = asyncHandler(async (req: CustomRequest, res: Response) => {
 
-    const goals = await GoalModel.find()
+    const goals = await GoalModel.find({ user: req.user?._id })
 
     res.status(200).json(goals)
 })
@@ -21,8 +26,13 @@ export const getGoals = asyncHandler(async (req: Request, res: Response) => {
     @access - Private
 
 */
-export const getGoal = asyncHandler(async (req: Request, res: Response) => {
-    res.status(200).json({ message: `Get goal ${req.params.id}` })
+export const getGoal = asyncHandler(async (req: CustomRequest, res: Response) => {
+    const goal = await GoalModel.find({ _id: req.params.id, user: req.user?._id })
+    if(!goal) {
+        res.status(400)
+        throw new Error("Goal not found")
+    }
+    res.status(200).json(goal)
 })
 
 /*
@@ -31,14 +41,15 @@ export const getGoal = asyncHandler(async (req: Request, res: Response) => {
     @access - Private
 
 */
-export const setGoal = asyncHandler(async (req: Request, res: Response) => {
+export const setGoal = asyncHandler(async (req: CustomRequest, res: Response) => {
     if(!req.body.text){
         // return res.status(400).json({message: "Please include a goal"})
         res.status(400)
         throw new Error("Please include a goal")   // this return an error html page to show u the details of the error.
     }
     const goal = await GoalModel.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user?._id
     })
 
     res.status(200).json(goal)
@@ -50,12 +61,26 @@ export const setGoal = asyncHandler(async (req: Request, res: Response) => {
     @access - Private
 
 */
-export const updateGoal = asyncHandler(async (req: Request, res: Response) => {
+export const updateGoal = asyncHandler(async (req: CustomRequest, res: Response) => {
 
     const goal = await GoalModel.findById(req.params.id)
     if(!goal) {
         res.status(400)
         throw new Error("Goal not found")
+    }
+
+    const user = await userModel.findById(req.user?.id)
+
+    // check for user
+    if(!user) {
+        res.status(401)
+        throw new Error("User not found")
+    }
+
+    // make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error("User not authorized")
     }
 
     const updatedGoal = await GoalModel.findByIdAndUpdate(req.params.id, req.body, { new: true})
@@ -69,7 +94,7 @@ export const updateGoal = asyncHandler(async (req: Request, res: Response) => {
     @access - Private
 
 */
-export const deleteGoal = asyncHandler(async (req: Request, res: Response) => {
+export const deleteGoal = asyncHandler(async (req: CustomRequest, res: Response) => {
 
     const goal = await GoalModel.findById(req.params.id)
     if (!goal) {
@@ -77,7 +102,22 @@ export const deleteGoal = asyncHandler(async (req: Request, res: Response) => {
         throw new Error("Goal not found")
     }
 
+    const user = await userModel.findById(req.user?.id)
+
+    // check for user
+    if (!user) {
+        res.status(401)
+        throw new Error("User not found")
+    }
+
+    // make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error("User not authorized")
+    }
+
+
     await GoalModel.deleteOne({ _id: req.params.id });
 
-    res.json({ id: req.params.id})
+    res.json({ id: req.params.id}) 
 })
